+++ deltafi-core-actions/src/main/java/org/deltafi/core/action/FilterByFiatTransformAction.java
@@ -1,0 +1,44 @@
+/*
+ *    DeltaFi - Data transformation and enrichment platform
+ *
+ *    Copyright 2022 DeltaFi Contributors <deltafi@deltafi.org>
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+
+package org.deltafi.core.action;
+
+import lombok.extern.slf4j.Slf4j;
+import org.deltafi.actionkit.action.Result;
+import org.deltafi.actionkit.action.filter.FilterResult;
+import org.deltafi.actionkit.action.transform.SimpleTransformAction;
+import org.deltafi.common.types.ActionContext;
+import org.deltafi.common.types.Content;
+import org.deltafi.common.types.SourceInfo;
+import org.jetbrains.annotations.NotNull;
+import org.springframework.stereotype.Component;
+
+import java.util.Map;
+
+@Component
+@Slf4j
+public class FilterByFiatTransformAction extends SimpleTransformAction {
+    public FilterByFiatTransformAction() {
+        super("Transform Action that always filters");
+    }
+
+    @Override
+    public Result transform(@NotNull ActionContext context, @NotNull SourceInfo sourceInfo, @NotNull Content content, @NotNull Map<String, String> metadata) {
+        return new FilterResult(context, "Filtered by fiat");
+    }
+}